fix: refactor cache path to use shared constants
