+++ packages/server/shared/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@activepieces/server-shared",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "dependencies": {
     "tslib": "^2.3.0",
     "@sentry/node": "7.120.0",
