+++ packages/server/shared/src/index.ts
@@ -13,3 +13,4 @@ export * from './lib/logger'
 export * from './lib/webhook-secrets-util'
 export * from './lib/exec'
 export * from './lib/system-props'
+export * from './lib/cache'
