+++ packages/server/worker/package.json
@@ -6,7 +6,7 @@
   "typings": "./src/index.d.ts",
   "dependencies": {
     "@activepieces/pieces-framework": "0.13.0",
-    "@activepieces/server-shared": "0.0.1",
+    "@activepieces/server-shared": "0.0.2",
     "@activepieces/shared": "0.12.0",
     "async-mutex": "0.4.0",
     "axios": "1.8.3",
@@ -18,7 +18,6 @@
     "nanoid": "3.3.8",
     "tslib": "2.6.2",
     "p-limit": "2.3.0",
-    "write-file-atomic": "5.0.1",
     "ws": "8.18.0",
     "tree-kill": "1.2.2"
   }
