+++ packages/server/worker/src/lib/piece-manager/development/pieces-builder.ts
@@ -1,8 +1,8 @@
 import { spawn } from 'child_process'
 import fs from 'fs/promises'
 import { Server } from 'http'
-import path, { resolve } from 'path'
-import { ApLock, filePiecesUtils, memoryLock, PiecesSource } from '@activepieces/server-shared'
+import { resolve } from 'path'
+import { ApLock, CacheState, filePiecesUtils, GLOBAL_CACHE_COMMON_PATH, memoryLock, PiecesSource } from '@activepieces/server-shared'
 import { debounce, isNil, WebsocketClientEvent } from '@activepieces/shared'
 import chalk from 'chalk'
 import chokidar from 'chokidar'
@@ -10,13 +10,6 @@ import { FastifyBaseLogger, FastifyInstance } from 'fastify'
 import { cacheState } from '../../cache/cache-state'
 
 export const PIECES_BUILDER_MUTEX_KEY = 'pieces-builder'
-
-const globalCachePath = path.resolve('cache', 'common')
-
-enum CacheState {
-    READY = 'READY',
-    PENDING = 'PENDING',
-}
 
 async function checkBuildTarget(nxProjectFilePath: string): Promise<string> {
     try {
@@ -68,7 +61,7 @@ async function handleFileChange(packages: string[], pieceProjectName: string, pi
 
         await filePiecesUtils(packages, log).clearPieceCache(piecePackageName)
 
-        const cache = cacheState(globalCachePath)
+        const cache = cacheState(GLOBAL_CACHE_COMMON_PATH)
         await cache.setCache('@activepieces/pieces-framework', CacheState.PENDING)
         await cache.setCache('@activepieces/pieces-common', CacheState.PENDING)
         await cache.setCache('@activepieces/shared', CacheState.PENDING)
