+++ packages/server/worker/src/lib/runner/process/factory/isolate-sandbox-process.ts
@@ -1,10 +1,9 @@
 import { spawn } from 'node:child_process'
 import path from 'node:path'
 import { arch } from 'node:process'
-import { execPromise, fileExists, PiecesSource } from '@activepieces/server-shared'
+import { execPromise, fileExists, GLOBAL_CACHE_COMMON_PATH, GLOBAL_CODE_CACHE_PATH, PiecesSource } from '@activepieces/server-shared'
 import { isNil } from '@activepieces/shared'
 import { FastifyBaseLogger } from 'fastify'
-import { GLOBAL_CACHE_COMMON_PATH, GLOBAL_CODE_CACHE_PATH } from '../../../cache/execution-files'
 import { workerMachine } from '../../../utils/machine'
 import { EngineProcess } from './engine-factory-types'
